Reestructure client components to reuse in the GUI
